pyproject.toml
[project]
name = "ai-job-recemmonder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.0.0",
    "fastapi>=0.116.1",
    "google-search-results>=2.4.2",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.31",
    "langchain-pinecone>=0.2.11",
    "numpy>=2.3.2",
    "pinecone-client>=6.0.0",
    "plotly>=6.3.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.35.0",
]
